+++ lib/settings/styles/timelineGroupStyle.d.ts
@@ -12,11 +12,19 @@ export interface TimelineGroupStyle {
     /**
      * Group stroke color.
      */
-    strokeColor: string;
+    strokeColor?: string;
+    /**
+     * Group stroke thickness.
+     */
+    strokeThickness?: number | null;
+    /**
+     * Group border radius. See canvas roundRect official documentation.
+     */
+    radii?: number | DOMPointInit | Iterable<number | DOMPointInit>;
     /**
      * Group fill color.
      */
-    fillColor: string;
+    fillColor?: string;
     /**
      * Group mouse over cursor style.
      */
@@ -28,5 +36,5 @@ export interface TimelineGroupStyle {
     /**
      * Style for all the keyframes in the current group.
      */
-    keyframesStyle: TimelineKeyframeStyle;
+    keyframesStyle?: TimelineKeyframeStyle;
 }